refactor: extract shared utilities from learn.ts into dedicated modules Move LearningObservation type/guards/parsing/formatting into observations.ts (pure data, no I/O), acquireMkdirLock into mkdir-lock.ts (generic lock primitive), and readObservations/writeObservations/warnIfInvalid/ updateDecisionsStatus into observation-io.ts (filesystem bridge). decisions.ts and legacy-decisions-purge.ts now import from the new modules instead of the duplicated/learn.ts copies. Test imports updated to match. Per ADR-001: no backward-compat re-exports. Co-Authored-By: Claude <noreply@anthropic.com>
